aimed-back-loginRedis/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pyjwt==2.9.0
redis==5.0.7
werkzeug==3.0.3
